Move keyDownListener to components
